requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-Mail==0.9.1
Flask-CORS==4.0.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
nltk==3.8.1
spacy==3.7.2
PyJWT==2.8.0
Werkzeug==3.0.1
